refactor: only support vue.js
