+++ src/adapter/element-ui/controls.js
@@ -1,5 +1,5 @@
 import omit from 'object.omit';
-import { cloneDeep } from '../../../utils.js';
+import { cloneDeep } from '../../utils.js';
 
 /**
  * Create a form control
