+++ src/adapter/index.js
